dev_requirements.txt
lxml-stubs
pylint==2.17.5
astroid==2.15.6
mypy==1.5.1
black==23.9.1
isort
types-cryptography
types-dataclasses
types-pycurl
types-python-dateutil
